pyproject.toml
[project]
name = "captcha-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.24.4",
    "pillow>=10.4.0",
    "scikit-image>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
ignore-vcs = true
packages = ["solver"]
exclude = ["__pycache__", "*.pyc", "*.exe~", "*.syso"]
